fix: reshuffle: moved old Graph content to RequestManagementSystem
