+++ src/DIRAC/RequestManagementSystem/private/test/RequestTaskTests.py
@@ -6,17 +6,7 @@
 
     test cases for RequestTask class
 """
-from __future__ import print_function
-from __future__ import absolute_import
-from __future__ import division
 
-__RCSID__ = "$Id $"
-# #
-# @file RequestTaskTests.py
-# @author Krzysztof.Ciba@NOSPAMgmail.com
-# @date 2013/03/27 15:59:40
-# @brief Definition of RequestTaskTests class.
-# # imports
 import unittest
 import importlib
 from mock import Mock, MagicMock
@@ -32,7 +22,7 @@ ReqClient = Mock(spec=ReqClient)
 from DIRAC.RequestManagementSystem.Client.Request import Request
 from DIRAC.RequestManagementSystem.Client.Operation import Operation
 
-########################################################################
+
 class RequestTaskTests(unittest.TestCase):
     """
     .. class:: RequestTaskTests
